python/requirements.txt
requests>=2.0
ultralytics>=8.1
tensorboard>=2.0
pillow>=9.0
